Replaced Item.getIcon() with Item.loadIcon()
loadIcon() was a void that did what it says but stored it in a global variable. To actually get the icon you had to call getIcon() which would then return the variable. This was fucking dumb so I made it so you just have to call getIcon() and it does both of those things. This breaks some things but I’m on mobile so